MB-67250: Fix narrowing warning in ep_testsuite
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/231029
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/tests/ep_test_apis.h
@@ -472,7 +472,7 @@ void wait_for_persisted_value(EngineIface* h,
 
 void wait_for_memory_usage_below(
         EngineIface* h,
-        int mem_threshold,
+        size_t mem_threshold,
         const std::chrono::seconds max_wait_time_in_secs = std::chrono::seconds{
                 60});
